Sửa một số lỗi sql và theo một số yêu cầu của sản xuất

+++ QLDuLieuTonKho_BTP/Models/DL_CD_Boc.cs
@@ -13,8 +13,8 @@ public class DL_CD_Boc
 
     [Required(ErrorMessage = "Ca không được để trống")]
     public string Ca { get; set; }
-    public double KhoiLuongTruocBoc { get; set; }
-    public double KhoiLuongPhe { get; set; } = 0;
+    public decimal KhoiLuongTruocBoc { get; set; }
+    public decimal KhoiLuongPhe { get; set; } = 0;
 
     [Required(ErrorMessage = "Người làm không được để trống")]
     public string NguoiLam { get; set; }
